Adjust expense Date dynamically
Expense dates and future expense dates are now calculated based on the time that the app is opened.

If the app is opened and it has passed payday, it will determine if expense has already passed. If so will mark it with a 3 flag and not include in upcoming expenses.